* master:
  Updated travis config
  Updated travis to use OS X image
  Version bump
  Clear derived data after running snapshot
  Drastically speeded up snapshot by not clearing derived data
  Revert "reimplemented grouped screenshots by orientation"
  Revert "fixed nil error because of invalid execution flow"
  Removed outdated comment
  Revert "updated helper template to add screnhot orientation to generated filename"
  Revert "Fixed style errors"
  Fixed style errors
  updated helper template to add screnhot orientation to generated filename
  fixed nil error because of invalid execution flow
  reimplemented grouped screenshots by orientation
  Enable support for iPhone 6s / iPhone 6s Plus and switch to sizes instead device names
  Fix Swift style

Conflicts:
	lib/assets/SnapshotHelper.swift

+++ lib/snapshot/runner.rb
@@ -44,6 +44,9 @@ module Snapshot
 
       # Generate HTML report
       ReportsGenerator.new.generate
+
+      # Clear the Derived Data
+      FileUtils.rm_rf(TestCommandGenerator.derived_data_path)
     end
 
     def config_launch_arguments
@@ -58,7 +61,8 @@ module Snapshot
 
     def launch(language, device_type, launch_arguments)
       screenshots_path = TestCommandGenerator.derived_data_path
-      FileUtils.rm_rf(screenshots_path)
+      FileUtils.rm_rf(File.join(screenshots_path, "Logs"))
+      FileUtils.rm_rf(screenshots_path) if Snapshot.config[:clean]
       FileUtils.mkdir_p(screenshots_path)
 
       File.write("/tmp/language.txt", language)
